testing item component with shallow rendering.
